# packages/api-contracts/openapi/backend.openapi.json
{
  "components": {
    "schemas": {
...
      "CreateOrganization": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "type": "object"
      },
...
      "EditOrganization": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "version"
        ],
        "type": "object"
      },
...
      "Failure": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "success": {
            "const": false,
            "type": "boolean"
          },
          "traceId": {
            "pattern": "^[0-9a-f]{32}$",
            "type": "string"
          }
        },
        "required": [
          "error",
          "success",
          "traceId"
        ],
        "type": "object"
      },
...
      "SuccessMutationResult": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MutationResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "traceId": {
            "pattern": "^[0-9a-f]{32}$",
            "type": "string"
          }
        },
        "required": [
          "data",
          "success",
          "traceId"
        ],
        "type": "object"
      },
      "SuccessOrganizationView": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganizationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "traceId": {
            "pattern": "^[0-9a-f]{32}$",
            "type": "string"
          }
        },
        "required": [
          "data",
          "success",
          "traceId"
        ],
        "type": "object"
      },
...
      "SuccessPageResponseOrganizationView": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PageResponseOrganizationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "const": true,
            "type": "boolean"
          },
          "traceId": {
            "pattern": "^[0-9a-f]{32}$",
            "type": "string"
          }
        },
        "required": [
          "data",
          "success",
          "traceId"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CustomerToken": {
        "description": "Bearer加一个空格再加原始opaque Token；仅客户小程序，无客户Cookie",
        "in": "header",
        "name": "X-Customer-Token",
        "type": "apiKey"
      },
      "PlatformCookie": {
        "description": "仅PLATFORM WEB；写请求必须独立CSRF及同源来源",
        "in": "cookie",
        "name": "__Secure-pet_platform_sid",
        "type": "apiKey"
      },
      "StaffCookie": {
        "description": "生产Cookie；local HTTP使用pet_dev_staff_sid，写请求另需X-CSRF-Token及固定Origin/Referer",
        "in": "cookie",
        "name": "__Secure-pet_staff_sid",
        "type": "apiKey"
      },
      "StaffToken": {
        "description": "Bearer加一个空格再加原始opaque Token；与Cookie不能并存",
        "in": "header",
        "name": "X-Staff-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "企业应用公共接口契约",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/identity/organizations": {
      "get": {
        "description": "identity:organization:list，TENANT范围；组织不决定业务数据范围。",
        "operationId": "listOrganizations",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessPageResponseOrganizationView"
                }
              }
            },
            "description": "成功，统一响应信封"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未知或格式错误输入"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或旧身份失效"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "无权限、保护账号、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "目标或关联不存在或不可访问"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本或生命周期冲突"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段验证失败"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障或写入结果未确认，禁止自动重试"
          }
        },
        "security": [
          {
            "StaffCookie": []
          },
          {
            "StaffToken": []
          }
        ],
        "summary": "本租户内部组织列表",
        "tags": [
          "identity-management-controller"
        ]
      },
      "post": {
        "operationId": "createOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMutationResult"
                }
              }
            },
            "description": "成功，统一响应信封"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未知或格式错误输入"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或旧身份失效"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "无权限、保护账号、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "目标或关联不存在或不可访问"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本或生命周期冲突"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段验证失败"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障或写入结果未确认，禁止自动重试"
          }
        },
        "security": [
          {
            "StaffCookie": []
          },
          {
            "StaffToken": []
          }
        ],
        "summary": "创建本租户内部组织",
        "tags": [
          "identity-management-controller"
        ]
      }
    },
    "/api/admin/identity/organizations/{id}": {
      "get": {
        "operationId": "getOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessOrganizationView"
                }
              }
            },
            "description": "成功，统一响应信封"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未知或格式错误输入"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或旧身份失效"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "无权限、保护账号、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "目标或关联不存在或不可访问"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本或生命周期冲突"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段验证失败"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障或写入结果未确认，禁止自动重试"
          }
        },
        "security": [
          {
            "StaffCookie": []
          },
          {
            "StaffToken": []
          }
        ],
        "summary": "本租户组织详情",
        "tags": [
          "identity-management-controller"
        ]
      },
      "put": {
        "operationId": "editOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditOrganization"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMutationResult"
                }
              }
            },
            "description": "成功，统一响应信封"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未知或格式错误输入"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或旧身份失效"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "无权限、保护账号、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "目标或关联不存在或不可访问"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本或生命周期冲突"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段验证失败"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障或写入结果未确认，禁止自动重试"
          }
        },
        "security": [
          {
            "StaffCookie": []
          },
          {
            "StaffToken": []
          }
        ],
        "summary": "修改本租户组织名称",
        "tags": [
          "identity-management-controller"
        ]
      }
    },
    "/api/admin/identity/organizations/{id}/status": {
      "put": {
        "operationId": "setOrganizationStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatus"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMutationResult"
                }
              }
            },
            "description": "成功，统一响应信封"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未知或格式错误输入"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或旧身份失效"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "无权限、保护账号、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "目标或关联不存在或不可访问"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本或生命周期冲突"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段验证失败"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障或写入结果未确认，禁止自动重试"
          }
        },
        "security": [
          {
            "StaffCookie": []
          },
          {
            "StaffToken": []
          }
        ],
        "summary": "启停本租户内部组织",
        "tags": [
          "identity-management-controller"
        ]
      }
    },
    "/api/admin/identity/permissions": {
      "get": {
        "description": "identity:role:detail；服务端计算可授予范围，写入仍独立检查。",
        "operationId": "listPermissionOptions",
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessListPermissionOption"
                }
              }
            },
            "description": "成功，统一响应信封"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未知或格式错误输入"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或旧身份失效"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "无权限、保护账号、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "目标或关联不存在或不可访问"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本或生命周期冲突"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段验证失败"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障或写入结果未确认，禁止自动重试"
          }
        },
        "security": [
          {
            "StaffCookie": []
          },
          {
            "StaffToken": []
          }
        ],
        "summary": "正式权限及可授予范围",
        "tags": [
          "identity-management-controller"
        ]
      }
    },
...
    "/api/admin/identity/roles/{roleId}/grants": {
      "put": {
        "description": "identity:role:grant；全量替换；同权限并集，不同权限不互借，受操作者逐权限可授予范围约束。",
        "operationId": "setRoleGrants",
        "parameters": [
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleGrants"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMutationResult"
                }
              }
            },
            "description": "成功，统一响应信封"
          },
          "400": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未知或格式错误输入"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "未登录或旧身份失效"
          },
          "403": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "无权限、保护账号、授予上限或CSRF拒绝"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "目标或关联不存在或不可访问"
          },
          "409": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "版本或生命周期冲突"
          },
          "422": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "字段验证失败"
          },
          "503": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Failure"
                }
              }
            },
            "description": "依赖故障或写入结果未确认，禁止自动重试"
          }
        },
        "security": [
          {
            "StaffCookie": []
          },
          {
            "StaffToken": []
          }
        ],
        "summary": "配置角色逐权限数据范围",
        "tags": [
          "identity-management-controller"
        ]
      }
    },
...
  }
}